refactor extract PR handler
