+++ app.js
@@ -2,10 +2,12 @@
 //
 // You installed the `dotenv` and `octokit` modules earlier. The `@octokit/webhooks` is a dependency of the `octokit` module, so you don't need to install it separately. The `fs` and `http` dependencies are built-in Node.js modules.
 import dotenv from "dotenv";
-import {App} from "octokit";
-import {createNodeMiddleware} from "@octokit/webhooks";
+import { App } from "octokit";
+import { createNodeMiddleware } from "@octokit/webhooks";
 import fs from "fs";
 import http from "http";
+import { handlePullRequestOpened } from "./handlers/pull-requests.js";
+
 
 // This reads your `.env` file and adds the variables from that file to the `process.env` object in Node.js.
 dotenv.config();
@@ -26,46 +28,8 @@ const app = new App({
 
 const config = JSON.parse(fs.readFileSync('config.json', 'utf8'));
 
-// This adds an event handler that your code will call later. When this event handler is called, it will log the event to the console. Then, it will use GitHub's REST API to add a comment to the pull request that triggered the event.
-async function handlePullRequestOpened({octokit, payload}) {
-  try {
-    // handle the event if the pull request is for a repository whose name matches the whitelist pattern in the config file but not the blacklist pattern
-    if (config.settings.whitelist && !config.settings.whitelist.some((pattern) => new RegExp(pattern).test(payload.repository.name))) {
-      console.log(`PR ignored because it's for a repository that does not match the whitelist pattern`);
-      return;
-    }
-    if (config.settings.exceptions && config.settings.exceptions.some((pattern) => new RegExp(pattern).test(payload.repository.name))) {
-      console.log(`PR ignored because it's for a repository that matches the blacklist pattern`);
-      return;
-    }
-
-    console.log(`Received a pull request event for #${payload.pull_request.number}`);
-    await postComment(octokit, payload);
-  } catch (error) {
-    if (error.response) {
-      console.error(`Error! Status: ${error.response.status}. Message: ${error.response.data.message}`)
-    }
-    console.error(error)
-  }
-};
-
-async function postComment(octokit, payload) {
-  // This defines the message that your app will post to pull requests.
-  const messageForNewPRs = "Thanks for opening a new PR! Please follow our contributing guidelines to make your PR easier to review.";
-
-  return await octokit.request("POST /repos/{owner}/{repo}/issues/{issue_number}/comments", {
-    owner: payload.repository.owner.login,
-    repo: payload.repository.name,
-    issue_number: payload.pull_request.number,
-    body: messageForNewPRs,
-    headers: {
-      "x-github-api-version": "2022-11-28",
-    },
-  });
-}
-
 // This sets up a webhook event listener. When your app receives a webhook event from GitHub with a `X-GitHub-Event` header value of `pull_request` and an `action` payload value of `opened`, it calls the `handlePullRequestOpened` event handler that is defined above.
-app.webhooks.on("pull_request.opened", handlePullRequestOpened);
+app.webhooks.on("pull_request.opened", (webhook) => handlePullRequestOpened(webhook, config));
 
 // This logs any errors that occur.
 app.webhooks.onError((error) => {
@@ -91,7 +55,7 @@ const localWebhookUrl = `http://${host}:${port}${path}`;
 //    - Check the signature of the incoming webhook event to make sure that it matches your webhook secret. This verifies that the incoming webhook event is a valid GitHub event.
 //    - Parse the webhook event payload and identify the type of event.
 //    - Trigger the corresponding webhook event handler.
-const middleware = createNodeMiddleware(app.webhooks, {path});
+const middleware = createNodeMiddleware(app.webhooks, { path });
 
 // This creates a Node.js server that listens for incoming HTTP requests (including webhook payloads from GitHub) on the specified port. When the server receives a request, it executes the `middleware` function that you defined earlier. Once the server is running, it logs messages to the console to indicate that it is listening.
 http.createServer(middleware).listen(port, () => {
